requirements.txt
google-generativeai
python-dotenv
ccxt
pandas-ta
fastapi
uvicorn
jinja2
black
flake8